document on manual labeling, add examples
