+++ README.md
@@ -4,7 +4,7 @@ Tool for analysis of security certificates and their security targets (Common Cr
 
 This project is developed by the [Centre for Research On Cryptography and Security](https://crocs.fi.muni.cz) at Faculty of Informatics, Masaryk University.
 
-## Usage (CC)
+## Install (CC)
 
 The tool requires several Python packages as well as the `pdftotext` binary somewhere on the `PATH`.
 The easiest way to setup the tool is to install it in a virtual environment, e.g.:
@@ -19,6 +19,12 @@ python3 -m venv virt
 . virt/bin/activate
 pip install -e .
 ```
+
+## Examples
+
+Some examples are documented in [examples](https://github.com/crocs-muni/sec-certs/blob/master/examples/)
+
+## Old API
 
 The following steps will do a full extraction and analysis of CC certificates:
